+++ tests/unit/test_clusterer_api/test_auto_clustering.py
@@ -1,3 +1,5 @@
+import time
+from relevanceai import Client
 from relevanceai.http_client import ClusterOps
 
 
@@ -17,3 +19,12 @@ def test_kmeans_clusterer(kmeans_clusterer: ClusterOps):
 
 def test_kmeans_cluster_report(kmeans_clusterer: ClusterOps):
     assert len(kmeans_clusterer.internal_report()) > 0
+
+
+def test_storing_reports(test_client: Client, kmeans_clusterer: ClusterOps):
+    report = kmeans_clusterer.internal_report()
+    result = test_client.store_cluster_report(report_name="simple", report=report)
+    time.sleep(2)
+    assert len(test_client.list_cluster_reports()) > 1
+    # Now delete the rep0ort - provide some cleanup :)
+    test_client.delete_cluster_report(result["_id"])